split main package into multiple files
